feat(ChargingCostService): Finalize Charges

+++ TeslaSolarCharger.Model/Entities/TeslaSolarCharger/HandledCharge.cs
@@ -5,8 +5,8 @@ public class HandledCharge
     public int Id { get; set; }
     public int ChargingProcessId { get; set; }
     public int CarId { get; set; }
-    public int? UsedGridEnergy { get; set; }
-    public int? UsedSolarEnergy { get; set; }
+    public decimal? UsedGridEnergy { get; set; }
+    public decimal? UsedSolarEnergy { get; set; }
     public decimal? CalculatedPrice { get; set; }
     public List<PowerDistribution> PowerDistributions { get; set; } = new();
